move progress indicator to common
